fix(Credits): stop rounding half-credit courses down to zero
NTUST issues 0.5-credit courses. Credits were an Int end to end, and both
parsers reached them through Int("0.5") / ^(\d+)$ — so every one of those
courses arrived as 0: missing from the class table's total, blank on the
transcript row, and weightless in the GPA.

- SDCourse.credits and CourseGrade.credits are Double; the transcript's
  credit patterns take an optional fraction, in every bracket form, so a
  half credit keeps its credit type instead of falling through to unknown
- CreditBreakdown follows, since the school's own summary totals add up
  the same halves
- `Double.creditsText` prints "3" and "0.5" rather than "3.0", locale-aware
- the cache DTO decodes older files unchanged — JSON has one number type —
  and the server already sent credits as a float, so the reconcile path
  stops truncating on the way in and on the way back out

Covered by HalfCreditTests: the transcript parser, the bracketed forms,
GPA weighting, and the formatter.

+++ swift/TigerDuck/Bridge/AppServiceBridge.swift
@@ -5,7 +5,7 @@ private struct CourseData: Sendable {
     let courseNo: String
     var courseName: String
     let instructor: String
-    let credits: Int
+    let credits: Double
     var classroom: String
     let enrolledCount: Int
     let maxCount: Int
@@ -296,7 +296,7 @@ enum AppServiceBridge {
                             courseName: c.courseName,
                             courseNameEn: nil,
                             moodleId: c.moodleIdNumber,
-                            credits: c.credits > 0 ? Double(c.credits) : nil,
+                            credits: c.credits > 0 ? c.credits : nil,
                             classroom: c.classroom.isEmpty ? nil : c.classroom,
                             instructors: c.instructor.isEmpty ? nil : [c.instructor],
                             scheduleJson: c.schedule.isEmpty ? nil : Dictionary(uniqueKeysWithValues: c.schedule.map { ("\($0.key)", $0.value) }),
@@ -666,7 +666,7 @@ enum AppServiceBridge {
             courseNo: first.CourseNo,
             courseName: first.CourseName,
             instructor: first.CourseTeacher,
-            credits: Int(first.CreditPoint) ?? 0,
+            credits: Double(first.CreditPoint) ?? 0,
             classroom: allClassrooms.joined(separator: ", "),
             enrolledCount: first.ChooseStudent ?? 0,
             maxCount: Int(first.Restrict2 ?? "0") ?? 0,
